Extract markdown sub/superscript parser to its own file.
